Still working on rebuilding the new model.. the application layer is still a mess. non of this works at the moment.
Repository is fleshed out, still needs to pull in revision logic from the application layer (getXrev..) and also the repository specific methods for svn such as checkout.

Revision is created and ready to be used, i think.

added last_remote_upadate to all new and existing tables to store when a row was last retrieved from the remote server.